Move AUTH_URL_GITHUB and AUTH_URL_GOOGLE to auth-services.js
